+++ src/reporters/pretty.ts
@@ -9,7 +9,7 @@ export const prettyFactory: IReporterFactory = {
 
 const center = 30;
 
-class PrettyReporter implements IReporter {
+export class PrettyReporter implements IReporter {
   private readonly numberFormat = Intl.NumberFormat('en-US', { maximumSignificantDigits: 3 });
   private readonly results: Benchmark[] = [];