+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## 1.1.0 2020-01-10
 
- - fix: allow the --grep option to be a real regex, not just substring
+- feat: allow async setup and teardown methods
+- refactor: tests and maintainability; add the ability to use matcha as a Node API
+- fix: allow the --grep option to be a real regex, not just substring
 
 ## 1.0.0 2019-12-22